requirements.txt
torch>=2.0
gymnasium>=0.29
scikit-learn>=1.3
matplotlib>=3.7
numpy>=1.24

python-dotenv>=1.0
tiktoken>=0.5
sentence-transformers>=2.2
datasets>=2.14

langchain-core>=0.3
langchain-openai>=0.3
langchain-google-genai>=2.0
langchain-anthropic>=0.3
langchain-huggingface>=0.1
langchain-ollama>=0.3       
